+++ src/browser/handleEvents.ts
@@ -40,12 +40,11 @@ const HandleEvents = {
   /**
    * 处理window的error的监听回到
    */
-  handleError(errorEvent: ErrorEvent): void {
+  handleError(errorEvent: ErrorEvent) {
     const target = errorEvent.target as ResourceErrorTarget
     if (target.localName) {
       // 资源加载错误 提取有用数据
       const data = resourceTransform(errorEvent.target as ResourceErrorTarget)
-      // push到行为栈
       breadcrumb.push({
         type: BREADCRUMBTYPES.RESOURCE,
         category: breadcrumb.getCategory(BREADCRUMBTYPES.RESOURCE),